# api/openapi.yaml
openapi: 3.0.3
info:
  title: Modura HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the Modura modular monolith.
servers:
  - url: /api
paths:
...
  /platform/auth/login:
    post:
      operationId: platformLogin
      summary: Establish a global platform-administrator session
      tags: [platform-authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformLoginRequest"
      responses:
        "200":
          description: Platform session established with a distinct refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
  /platform/auth/refresh:
    post:
      operationId: platformRefresh
      summary: Rotate a platform-administrator refresh secret
      tags: [platform-authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Platform session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate the refresh secret and issue an access token
      tags: [authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
  /organization/departments:
    get:
      operationId: listDepartments
      summary: List departments in the authenticated tenant
      tags: [organization]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant department list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
    post:
      operationId: createDepartment
      summary: Create a department in the authenticated tenant
      tags: [organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierResponse"
        "400":
          description: Invalid department input
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
  /organization/departments/{departmentId}:
    parameters:
      - $ref: "#/components/parameters/DepartmentId"
    patch:
      operationId: moveDepartment
      summary: Move a department under another parent
      tags: [organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveDepartmentRequest"
      responses:
        "204":
          description: Department moved
        "400":
          description: Move violates a department invariant
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
        "404":
          description: Department or parent not found in the tenant
    delete:
      operationId: deleteDepartment
      summary: Delete an unused non-root department
      tags: [organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Department deleted
        "400":
          description: Root or referenced department cannot be deleted
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
        "404":
          description: Department not found in the tenant
  /organization/positions:
    get:
      operationId: listPositions
      summary: List positions in the authenticated tenant
      tags: [organization]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant position list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
    post:
      operationId: createPosition
      summary: Create a position in the authenticated tenant
      tags: [organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionRequest"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierResponse"
        "400":
          description: Invalid or duplicate position
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
  /organization/users/{userId}/assignment:
    put:
      operationId: assignUserOrganization
      summary: Set a user's primary department and optional position
      tags: [organization]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignUserOrganizationRequest"
      responses:
        "204":
          description: User organization assignment replaced
        "400":
          description: Invalid or cross-tenant assignment
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/AuthorizationFailed"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
    DepartmentId:
      name: departmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthorizationFailed:
      description: The authenticated actor lacks the required permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    PlatformLoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
...
    Department:
      type: object
      additionalProperties: false
      required: [id, name, sortOrder]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        sortOrder:
          type: integer
    CreateDepartmentRequest:
      type: object
      additionalProperties: false
      required: [parentId, name, sortOrder]
      properties:
        parentId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 128
        sortOrder:
          type: integer
          minimum: -2147483648
          maximum: 2147483647
    MoveDepartmentRequest:
      type: object
      additionalProperties: false
      required: [parentId]
      properties:
        parentId:
          type: string
          format: uuid
    IdentifierResponse:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id:
          type: string
          format: uuid
    Position:
      type: object
      additionalProperties: false
      required: [id, name, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, disabled]
    CreatePositionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
    AssignUserOrganizationRequest:
      type: object
      additionalProperties: false
      required: [departmentId]
      properties:
        departmentId:
          type: string
          format: uuid
        positionId:
          type: string
          format: uuid
          nullable: true
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, csrfToken]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        csrfToken:
          type: string
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string